feat(medication): EPI-1129: Automatically discontinue medication when end date and time reached
