+++ packages/facility-server/app/tasks/index.js
@@ -3,11 +3,9 @@ import config from 'config';
 import { SendStatusToMetaServer } from '@tamanu/shared/tasks/SendStatusToMetaServer';
 
 // import { SenaitePoller } from './SenaitePoller';
-import { MedicationDiscontinuer } from './MedicationDiscontinuer';
 import { RefreshUpcomingVaccinations } from './RefreshMaterializedView';
 
 const DEFAULT_TASK_CLASSES = [
-  MedicationDiscontinuer,
   RefreshUpcomingVaccinations,
   SendStatusToMetaServer,
 ];
