+++ packages/facility-server/config/default.json5
@@ -106,10 +106,6 @@
     "requestFailureRate": 0
   },
   "schedules": {
-    "medicationDiscontinuer": {
-      // every day at 12:01 AM
-      "schedule": "1 0 * * *"
-    },
     "refreshMaterializedView": {
       "upcomingVaccinations": {
         // every 10 mins
